Refactor role checks in middleware classes

See merge request shasank7600/education_management2!32

+++ database/migrations/0001_01_01_000008_create_staff_information.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreign('institute_id')->references('institute_id')->on('institute')->onDelete('cascade');
             $table->string('staff_name',50);
             $table->enum('gender',["M","F","O"]);
-            $table->integer('contact_number');
+            $table->unsignedBigInteger('contact_number');
             $table->string('email');
             $table->text('address');
             $table->string('designation');
